: Fix nullable Activity route not working for bigger Activity
For some reason, on iOS it was crashing when navigating to Edit an activity if it was big enough. Not exactly clear, but it was due to the activity being nullable. If it was not nullable, it would work like Android

Also changed the archive selection to only get the activity IDs since it was not needed to pass the whole thing