Feature: KMP & iOS app (#6)
* Migrate to Koin to prepare for KMP

* Update some dependencies to latest versions

* Migrate to KMP Compose Navigation and to KT DateTime

* Update to Kotlin 2.2 and to KSP 2

* Remove Accompanist

* Migrate Firebase to KMP variants

* Add iOS App

* : Initial major KMP transition

Migrate from regular resources to Compose Resources
Add AuthUiHandler to handle Firebase Auth UI in Android/iOS
WIP: Add Firebase to iOS + CocoaPods
WIP: Add back Pathfinders/Unity flavour

* Remove some unused libs and try to fix iOS build

* iOS app compiling (but not launching)

* iOS app compiling and launching (but crashing)

* Fix Admin list functions

* Auth UI working on both Android & iOS

* Fix more screens in iOS

* Add back call action

* : Remove Firebase DI providers

They do not seem compatible due to expect/actual, so we can't inject them through constructors

* : Fix nullable Activity route not working for bigger Activity

For some reason, on iOS it was crashing when navigating to Edit an activity if it was big enough. Not exactly clear, but it was due to the activity being nullable. If it was not nullable, it would work like Android

Also changed the archive selection to only get the activity IDs since it was not needed to pass the whole thing

* Add back Lottie animation using Compottie

* Remove unnecessary TODO

* Initial flavouring for iOS

* Add icon and name to Unit iOS

* Add back dynamic colors for Android and fix AuthUI edge-to-edge

* Add back flavour-based participant class

* Add back app logo to home screen

* : Add back color schemes based on flavour

Fix drawable for both flavours

* Proper email method, fix Google SignIn on iOS

* Fix release variant compilation for iOS

* Share version names and codes between Android & iOS

* Fix ktlint and fix lint errors

* Revert task configuration

* Properly fix ktlint

* Small version updates

* Better way to interact with dates and small fixes